Start fixing tests
