+++ test/test_obstacles.test
@@ -1,0 +1,3 @@
+<launch>
+  <test test-name="test_obstacles" pkg="local_pathfinding" type="test_obstacles.py" />
+</launch>
